classes/LiveStock.ts
class Livestock {
  id: string;
  healthStatus: string;
  location: string;
  dietPlan: string;

  constructor(
    id: string,
    healthStatus: string,
    location: string,
    dietPlan: string
  ) {
    this.id = id;
    this.healthStatus = healthStatus;
    this.location = location;
    this.dietPlan = dietPlan;
  }

  trackLocation(rfidData: any) {
    // Logic to update the location using RFID data
    this.location = rfidData.location;
  }

  monitorHealth() {
    // Logic to monitor health status
    // This could involve checking sensors, health reports, etc.
    console.log(`Monitoring health status of livestock ${this.id}`);
  }

  feed() {
    // Logic to apply the diet plan
    console.log(
      `Feeding livestock ${this.id} with diet plan: ${this.dietPlan}`
    );
  }
}

export { Livestock };
